+++ packages/contact-center/cc-components/src/components/task/task.types.ts
@@ -551,3 +551,33 @@ export interface AutoWrapupTimerProps {
   allowCancelAutoWrapup?: boolean;
   handleCancelWrapup: () => void;
 }
+
+export interface CallControlButton {
+  id: string;
+  icon: string;
+  onClick?: () => void;
+  tooltip: string;
+  className: string;
+  disabled: boolean;
+  isVisible: boolean;
+  menuType?: CallControlMenuType;
+}
+
+export interface ControlVisibility {
+  accept: boolean;
+  decline: boolean;
+  end: boolean;
+  muteUnmute: boolean;
+  holdResume: boolean;
+  consult: boolean;
+  transfer: boolean;
+  conference: boolean;
+  wrapup: boolean;
+  pauseResumeRecording: boolean;
+  endConsult: boolean;
+  recordingIndicator: boolean;
+}
+
+export interface MediaTypeInfo {
+  labelName: string;
+}
